FEAT Order 기능 구현 및 QueryDSL 세팅 및 적용

+++ src/main/java/com/example/deliverybackend/interfaces/restaurant/RestaurantDto.java
@@ -25,7 +25,7 @@ public class RestaurantDto {
 //        @NotBlank
         private Double longitude;
 
-        private Double deliveryFee;
+        private Integer deliveryFee;
 
         public Point setPoint() {
             return GeometryPoint.createPoint(latitude, longitude);
